# OpenApi/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/Accounts/register": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "operationId": "Accounts_RegisterUser",
        "requestBody": {
          "x-name": "userForRegistration",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserForRegistrationDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Accounts/login": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "operationId": "Accounts_Login",
        "requestBody": {
          "x-name": "userForAuthentication",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserForAuthenticationDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserForRegistrationDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "confirmPassword": {
            "type": "string"
          }
        }
      },
      "UserForAuthenticationDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
    }
  }
}